OBGM-47 Refactor ProductSupplierAttributeExcelImporter
